Android 15.0.0 Release 20 (BP1A.250305.019)

# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCZ8epTgAKCRDorT+BmrEO
# eB59AJ91rMk2cTwTy3p3F4MjkF7XEORLJQCff5O1RGa2khd8+JlrZ7TVN54hRXo=
# =iJYQ
# -----END PGP SIGNATURE-----
# gpg: Signature made Wed Mar  5 03:30:54 2025 EET
# gpg:                using DSA key 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Good signature from "The Android Open Source Project <initial-contribution@android.com>" [ultimate]

* tag 'android-15.0.0_r20': (412 commits)
  Undo NAN periodic ranging changes in the legacy HAL and AIDL conversion method.
  Update defaults for IVehicleGeneratedHeaders-V3.
  Add AIDL API feedback
  Add check volume group id uniqueness
  Added VEHICLE_PASSIVE_SUSPENSION_HEIGHT to HAL
  Added BRAKE_FLUID_LEVEL_LOW to HAL
  Added BRAKE_PAD_WEAR_PERCENTAGE to HAL
  Added BRAKE_PEDAL_COMPRESSION_PERCENTAGE to HAL
  Added ACCELERATOR_PEDAL_COMPRESSION_PERCENTAGE to HAL
  Added VEHICLE_DRIVING_AUTOMATION_TARGET_LEVEL to HAL
  Added INSTANTANEOUS_EV_EFFICIENCY to HAL
  Added INSTANTANEOUS_FUEL_ECONOMY to HAL
  Added VEHICLE_HORN_ENGAGED to HAL
  Import contexthub v4 from bluetooth socket hal
  Added SAE_J3400 enums to EvConnectorType.aidl
  Added TURN_SIGNAL_LIGHT_STATE and TURN_SIGNAL_SWITCH to HAL
  Revert "audio: update setAudioPatch"
  Remove ValueRange.
  Add min_sdk_version to vehicle HAL property protos Rust library
  audio: update setAudioPatch
  ...

+++ graphics/composer/2.2/utils/command-buffer/include/composer-command-buffer/2.2/ComposerCommandBuffer.h
@@ -17,18 +17,12 @@
 #pragma once
 
 #ifndef LOG_TAG
-#warn "ComposerCommandBuffer.h included without LOG_TAG"
+#warning "ComposerCommandBuffer.h included without LOG_TAG"
 #endif
 
 //#define LOG_NDEBUG 0
 
-#include <algorithm>
-#include <limits>
-#include <memory>
 #include <vector>
-
-#include <inttypes.h>
-#include <string.h>
 
 #include <android/hardware/graphics/composer/2.2/IComposer.h>
 #include <android/hardware/graphics/composer/2.2/IComposerClient.h>
@@ -82,7 +76,7 @@ class CommandWriterBase : public V2_1::CommandWriterBase {
 
     void setLayerPerFrameMetadata(const hidl_vec<IComposerClient::PerFrameMetadata>& metadataVec) {
         beginCommand(IComposerClient::Command::SET_LAYER_PER_FRAME_METADATA,
-                     metadataVec.size() * 2);
+                     static_cast<uint16_t>(metadataVec.size() * 2));
         for (const auto& metadata : metadataVec) {
             writeSigned(static_cast<int32_t>(metadata.key));
             writeFloat(metadata.value);
